third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger-ui/reveal-not-skipped.html

<html>
<head>
<script src="../../../inspector/inspector-test.js"></script>
<script src="../../../inspector/debugger-test.js"></script>
<script src="../debugger/resources/script1.js"></script>
<script>
function throwAnException()
{
    var i = 0; var j = i + 2; var k = j + i;

    return unknown_var;
}

var test = function() {
  var panel = UI.panels.sources;
  TestRunner.DebuggerAgent.setPauseOnExceptions(SDK.DebuggerModel.PauseOnExceptionsState.PauseOnUncaughtExceptions);

  SourcesTestRunner.runDebuggerTestSuite([
    function testRevealAfterPausedOnException(next) {
      TestRunner.addResult('Showing script1 source...');
      SourcesTestRunner.showScriptSource('script1.js', step2);

      function step2() {
        TestRunner.addResult('Script source was shown for \'' + panel.visibleView._uiSourceCode.name() + '\'.');
        TestRunner.addResult('Throwing exception...');
        TestRunner.evaluateInPage('setTimeout(throwAnException, 0)');
        TestRunner.addSniffer(Sources.TabbedEditorContainer.prototype, 'showFile', step3);
      }

      function step3() {
        TestRunner.addResult('Script source was shown for \'' + panel.visibleView._uiSourceCode.name() + '\'.');
        TestRunner.addResult('Reloading page...');
        TestRunner.reloadPage(step4);
      }

      function step4() {
        TestRunner.addResult('Showing script1 source...');
        SourcesTestRunner.showScriptSource('script1.js', step5);
      }

      function step5() {
        TestRunner.addResult('Script source was shown for \'' + panel.visibleView._uiSourceCode.name() + '\'.');
        TestRunner.addResult('Throwing exception...');
        TestRunner.evaluateInPage('setTimeout(throwAnException, 0)');
        TestRunner.addSniffer(Sources.TabbedEditorContainer.prototype, 'showFile', step6);
      }

      function step6() {
        TestRunner.addResult('Script source was shown for \'' + panel.visibleView._uiSourceCode.name() + '\'.');
        next();
      }
    },

    function testRevealAfterPrettyPrintWhenPaused(next) {
      TestRunner.addResult('Throwing exception...');
      SourcesTestRunner.waitUntilPaused(step2);
      function step2() {
        TestRunner.addResult('Showing script1 source...');
        SourcesTestRunner.showScriptSource('script1.js', step3);
      }

      function step3() {
        TestRunner.addResult('Script source was shown for \'' + panel.visibleView._uiSourceCode.name() + '\'.');
        TestRunner.addResult('Formatting...');
        SourcesTestRunner.scriptFormatter().then(function(scriptFormatter) {
          TestRunner.addSniffer(
              Sources.ScriptFormatterEditorAction.prototype, '_updateButton', uiSourceCodeScriptFormatted);
          scriptFormatter._toggleFormatScriptSource();
        });
      }

      function uiSourceCodeScriptFormatted() {
        TestRunner.addResult('Script source was shown for \'' + panel.visibleView._uiSourceCode.name() + '\'.');
        next();
      }
    }
  ]);
};
</script>
</head>
<body onload="runTest()">
<p>Tests that certain user actions in scripts panel reveal execution line.</p>
</body>
</html>
